[dev.unified] cmd/compile/internal/noder: add optExpr for optional expressions
Previously, {writer,reader}.expr would allow for nil
expressions (i.e., no expression at all, not a "nil" identifier). But
only a few contexts allow this, and it simplifies some logic if we can
assume the expression is non-nil.

So this CL introduces optExpr as a wrapper method for handling nil
expressions specially.

Change-Id: I438bae7a3191126f7790ec0bf5b77320fe855514