cmd/compile: enable jump tables for PPC64
Use arm64 as a template. The implentation is slightly more verbose
as PPC64 requires a few extra instructions.

Similarly, we use CTR to perform indirect jumps, this requires
marking the mtctr + bctr sequences as unsafe. LR is not desirable
as it will require all functions which use jump tables to stack
a frame.

Change-Id: I772dd8198afd2bc4c7774798c37f1bec59e186f0