cmd/go: clean modroot before scan walk to enable prefix trim
We use str.TrimFilePathPrefix to trim the module root prefix and get the
relative path of each package in the module when scanning the module.
Make sure to clean the path before doing the walk to ensure that each
path starts with that prefix, because walk will clean the root path before
joining each subdirectory path to it.

Change-Id: I1dc1eddbd42030eb6d5d8e76a8675f94216447c3