cmd/compile: add new ssa pass csedown
DO NOT SUBMIT
DO NOT REVIEW

This pass enables to eliminate values that will endup being
reused together.

In other words it do common-subexpression elimination downward in
the control flow.

For example it rewrites:
  if c {
    x /= -y
  } else {
    x /= y
  }

In:
  var t int
  if c {
    t = -y
  } else {
    t = y
  }
  x /= t

The goal of this pass is not gain performance.
But to cleanup behind opt matching through phis (not implemented yet).

Updates #53300

Change-Id: I58118e21b001d123d1f514ab0d8449b061f514fe