cmd/compile: experimental loop iterator capture semantics change
DO NOT SUBMIT

For loops of the form `for x,y := range thing`, if x and/or y
are addressed or captured by a closure, are transformed by
renaming x/y to a temporary and prepending an assignment to
the body of the loop x := tmp_x.  This changes the loop semantics
by making each iteration's instance of x be distinct from the
others (currently they are all aliased, and when this matters,
it is almost always a bug).

3-range with captured iteration variables are also transformed,
though it is a more complex transformation.

Change-Id: I180a51a3f8d4173f6210c861f10de23de8a1b1db