go/token: use atomics not Mutex for last file cache
Previously, FileSet would cache the last *File found by a lookup,
using a full (exclusive) mutex within FileSet.File, turning a logical
read operation into an update. This was one of the largest sources
of contention in gopls.

This change uses atomic load/store on the 'last' field without
a mutex. It is crucial that the s.last=f operation happens
outside the critical section that creates f to ensure that all
assignments to the fields of f happen before f is published
to another thread.

Change-Id: Ice8641650d8495b25b0428e9b9320837ff2ca7e1