html: improve EscapeString and UnescapeString performance using avx instructions on _amd64
The existing implementation does not take advantage of SIMD or avx, and could benefit from
from an optimized implementaiton. Embedded hash tables are used to replace entity and entity2.
A unit test was added to validate the new hash is functionally identical to the entity and entity2 maps.

Addresses proposal #53410 (https://github.com/golang/go/issues/53410)

Change-Id: Ia3a8152644bd341ae3e8f33b4ec3f9b42ac3f6ca