cmd/compile: only skip visit closures in generic function call
CL 412794 makes subster skip substituting closures when processing
unsafe builtins argument. However, it's too agressive, since when we
only care about the case the closure belong to body of a generic
function call, because the function call is discarded, so we lack of
derived type information, so we can't instantiate the closure.

For other expressions, we always have the derived types info, so it's ok
to substitutie the closure.

Fixes #53439

Change-Id: I6e83927f8b453e434a71665c39e2134c6e1fdc7e