spec: add section on reference types
Add sentences to slice, map, and channel types
with a link to the new section.

Also, remove notion of "passing by value" in section on
calls: it is not needed for Go as paramater passing is
exactly the same as assignment.

This change copies some prose from CL 391634 by Alan Donovan.

Fixes #5083.

Change-Id: Ia11e9d61bc084a9df8adf2c5a3290c67938e3019