cmd/internal/obj/mips,s390x: save LR after decrementing SP
Following CL 412474, for the rest of the LR architectures. On
MIPS(32/64) and S390X, there is no single instruction that saves
the LR and decrements the SP, so we need to insert an instruction
to save the LR after decrementing the SP.

On ARM(32) and PPC64 we already use a single instruction to save
the LR and decrement the SP.

On RISCV we use two instructions, but I haven't been able to
reproduce the failure. So we'll leave it out for now.

Updates #53374.

Change-Id: I5a2e211026d95edb0e0f7d084ddb784f8077b86d