os/exec: LookPath: use access for exec check
Having an executable bit set for a binary is not enough for it to be
executable -- there might be more checks in the kernel. For example,
binaries on a filesystem mounted with "noexec" flag couldn't be
executed. There might be other scenarios involving ACLs, SELinux,
capabilities, and so on. Such scenarios also require the check for
for ability to execute via access(2) syscall with X_OK flag (alas,
we still have to use stat(2) to filter out directories).

As a result (of not using access), LookPath might either find a
non-executable (while going over $PATH elements), or return a false
positive that the argument provided is an executable.

Fix this, and add a test case (which, unfortunately, requires root,
and is currently Linux-specific, since different platforms either
have different semantics for tmpfs with noexec, or lack the feature).

Change-Id: If49b6ef6bf4dd23b2c32bebec8832d83e511a4bb