testing: give an example for ReportMetrics in a parallel benchmark
This change adds an example of using ReportMetrics in a parallel benchmark, specifically indicating the allowed usage of b.N as the iteration count. It intends to point out that the contents of the for loop in a parallel benchmark is executed concurrently, meaning race conditions can occur and need to be accounted for. It also attempts to indicate the behavior of the benchmarking function respective to b.N as to give a hint as to why it shouldn't be relied upon for the iteration count within the benchmark for loop body. The output of the benchmark is something like this depending on the machine:

```
...
Running the benchmark 1 times...
Running the benchmark 100 times...
Running the benchmark 10000 times...
Running the benchmark 1000000 times...
Running the benchmark 5674234 times...
BenchmarkName-GOMAXPROCS    	 5674234	       220.6 ns/op	        10.00 compares/op
```

Fixes #50756

Change-Id: Icccb3ed1503704e554744defb2163dcd2498db60
GitHub-Last-Rev: 9bdeb0a301b0db55cbea0f2f25cd94de68822366
GitHub-Pull-Request: golang/go#53603