go/ast: record comments precisely in the syntax tree
Since its design over a decade ago, the go/ast package
has recorded functional syntax as a tree, and comments
as a separate list in the File. (A few comments were
handled specially in the tree, such as doc comments and
end-of-line comments.) This approach was simple to implement
as it avoided both the space overheads of storing all
comments in the tree, and the tricky question of which
node should own each comment. However, it places extra
burdens on the printer, which had to interleave the
comment list with the syntax tree during the traversal,
based on position information in the comment. Also,
refactoring tools that delete, move, or insert syntax
subtrees have to go to great lengths to ensure that
associated comments are deleted or moved correctly.

It has been a major obstacle to the development of such
tools, and has spurred alternative representations of
syntax such as the excellent github.com/dave/dst.
However, we believe the standard library can do better.

This CL is an early proof of concept of a new representation
of comments within the syntax tree that faithfully preserves
the relative order of tokens and comments, allowing a
printer to emit the same token stream without consulting
position information. A rough demonstration of the printer is
included as a test of the parser, but a complete implementation
at the quality of go/printer is a larger project.
We believe this suggests the representation is complete,
but it is not yet convenient. Further work is needed to experiment
with refactoring tools based on the representation to see what
additional operations the package should support.

The representation is based on two principles:

1) comments are rare (2.5% of nodes), and not always wanted,
   yet the set of places they may appear is numerous.
   For example, a var declaration may have comments in
   four places:  .var.(specs.).
   Each period is a potential comment attachment point
   that maps to a list of comments.
   We cannot afford each of these to have its own field
   in the node struct. Each field may use at most one
   typically-nil pointer to represent all its immediate
   comments. This leads to a representation of the
   mapping from points to comments as a single slice
   of (point, comment) pairs.

2) the invariant that every Decl, Stmt, Expr, and Type is
   capable of holding comments at its start and end;
   this alleviates the parent node from the responsibility
   to hold comments before or after any of these trees.
   For example, a BinaryExpr   x + y   need not hold
   any additional comments, because all four places in
   which they might appear .X.+.Y. are accounted for
   by the subexpressions. By contrast, a UnaryExpr  - x
   must be able to hold comments appearing before the
   operator symbol. This reduces the number of node objects
   that require new fields by about 26%.

WORK IN PROGRESS
TODO
- API documentation
- benchmarks of space usage:
  - +17% heap overhead when loading std with go/packages LoadSyntax
    even when precise comments are disabled. This is the cost of
    extra fields in structs, in particular Ident and BasicLit,
    which push into a higher runtime size class.
  - a further 1.2% when the feature is actually enabled, suggesting
    that it should be always on.
  - +3.4% if the optimization in principle 2 is ignored.
- benchmarks of CPU when disabled and enabled
- documentation of the representation, parser changes,
  printer proof-of-concept.
- helper functions for refactoring tools, e.g.
  replace tree X with tree Y, taking/leaving comments.
- design of Space feature, not yet implemented, to record quality
  of intercomment whitespace.
- experiment with refactoring tools.
- Questions:
  - is the space cost acceptable?
  - should it be always on?
  - naming

DO NOT REVIEW -- needs a proposal
DO NOT SUBMIT

Fixes golang/go#20744

Change-Id: I660530a69472f0c8d5837c7d46ddde2ad1e8b282