cmd/internal/obj/loong64: save LR after decrementing SP
Refer to CL 413428 and 412474, for Loong64, like mips, s390x and riscv there
is no single instruction that saves the LR and decrements the SP, so we also
need to insert an instruction to save the LR after decrementing the SP. This
will lead to the exception of the instruction to be executed due to the use
of the contents below the SP that may have been damaged.

Fixes #56623.
Updates #53374.

Change-Id: I3de040792f0a041d3d2a98ea89c23a2dd2f4ad10